IB::Contract has a method: read_contract_from tws; default-parameters in contract_config.yml
In the modul tws_reader, three methods are defined:
First a wrapper for the default-attributes (ib-ruby/contract_config.yml)
Second, update_query, which extracts  nessesary attributes from the contract_record in order to perform a tws-request
and finally »read_contract_from_tws« which updates the record with data from the TWS.
If the tws-query is not successfull, con_id is set to zero and a short message ist returned via local_symbol.
If the contract lives in the database, the contract_detail-Record is created / updated as well.

A valid connection to the tws is required to perform the query.