Model: Contract – skipping validation and default-value
In Order to save a IB::Contract to the DB, that can be used to query the TWS, only the con_id and the currency-Attributes are ressesary. In previous Versions, the default-Exchange is 'smart'. This is assigned when saved. 
It was not possible to save a IB::Contract without a sec_type either.

To do: Extend IB:Contract + IB:Option with IB::Stock, IB:Future, IB:Forex, IB:Bond